+++ tf2/include/tf2/buffer_core.h
@@ -91,7 +91,7 @@ static constexpr Duration BUFFER_CORE_DEFAULT_CACHE_TIME = std::chrono::seconds(
  *
  * All function calls which pass frame ids can potentially throw the exception tf::LookupException
  */
-class BufferCore : public BufferCoreInterface
+class BufferCore : virtual public BufferCoreInterface
 {
 public:
   /************* Constants ***********************/
@@ -144,22 +144,6 @@ public:
     const std::string & source_frame,
     const tf2::TimePoint & time) const override;
 
-#if !TF2_ROS_FREE_CORE
-  /** \brief Get the transform between two frames by frame ID.
-   * \param target_frame The frame to which data should be transformed
-   * \param source_frame The frame where the data originated
-   * \param time The time at which the value of the transform is desired. (0 will get the latest)
-   * \return The transform between the frames as a ROS type.
-   *
-   * Possible exceptions tf2::LookupException, tf2::ConnectivityException,
-   * tf2::ExtrapolationException, tf2::InvalidArgumentException
-   */
-  TF2_PUBLIC
-  geometry_msgs::msg::TransformStamped
-  lookupTransform(
-    const std::string & target_frame, const std::string & source_frame,
-    const TimePoint & time) const override;
-#endif
 
   /**
    * \brief Get the transform between two frames by frame ID assuming fixed frame.
@@ -182,27 +166,6 @@ public:
     const std::string & source_frame,
     const tf2::TimePoint & source_time,
     const std::string & fixed_frame) const override;
-
-#if !TF2_ROS_FREE_CORE
-  /** \brief Get the transform between two frames by frame ID assuming fixed frame.
-   * \param target_frame The frame to which data should be transformed
-   * \param target_time The time to which the data should be transformed. (0 will get the latest)
-   * \param source_frame The frame where the data originated
-   * \param source_time The time at which the source_frame should be evaluated. (0 will get the latest)
-   * \param fixed_frame The frame in which to assume the transform is constant in time.
-   * \return The transform between the frames as a ROS type.
-   *
-   * Possible exceptions tf2::LookupException, tf2::ConnectivityException,
-   * tf2::ExtrapolationException, tf2::InvalidArgumentException
-   */
-
-  TF2_PUBLIC
-  geometry_msgs::msg::TransformStamped
-  lookupTransform(
-    const std::string & target_frame, const TimePoint & target_time,
-    const std::string & source_frame, const TimePoint & source_time,
-    const std::string & fixed_frame) const override;
-#endif
 
   TF2_PUBLIC
   tf2::Stamped<std::pair<tf2::Vector3, tf2::Vector3>> lookupVelocityTf2(
