+++ tf2_py/package.xml
@@ -17,7 +17,7 @@
 
   <build_depend>geometry_msgs</build_depend>
 
-  <depend>tf2</depend>
+  <depend>tf2_ros</depend>
 
   <exec_depend>builtin_interfaces</exec_depend>
   <exec_depend>geometry_msgs</exec_depend>
