+++ tf2_py/src/tf2_py.cpp
@@ -3,6 +3,8 @@
 #include <tf2/buffer_core.h>
 #include <tf2/exceptions.h>
 
+#include <tf2_ros/buffer.h>
+
 #include <string>
 #include <vector>
 
@@ -105,7 +107,7 @@ static PyObject * tf2_connectivityexception = nullptr, * tf2_lookupexception = n
 struct buffer_core_t
 {
   PyObject_HEAD
-  tf2::BufferCore * bc;
+  tf2_ros::BufferCoreROSConversions * bc;
 };
 
 static PyObject * transform_converter(const geometry_msgs::msg::TransformStamped * transform)
@@ -381,7 +383,7 @@ static int BufferCore_init(PyObject * self, PyObject * args, PyObject * kw)
     return -1;
   }
 
-  reinterpret_cast<buffer_core_t *>(self)->bc = new tf2::BufferCore(cache_time);
+  reinterpret_cast<buffer_core_t *>(self)->bc = new tf2_ros::BufferCoreROSConversions(cache_time);
 
   return 0;
 }
@@ -404,20 +406,20 @@ static void BufferCore_finalize(PyObject * self)
 static PyObject * allFramesAsYAML(PyObject * self, PyObject * args)
 {
   (void)args;
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   return stringToPython(bc->allFramesAsYAML());
 }
 
 static PyObject * allFramesAsString(PyObject * self, PyObject * args)
 {
   (void)args;
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   return stringToPython(bc->allFramesAsString());
 }
 
 static PyObject * canTransformCore(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * target_frame, * source_frame;
   tf2::TimePoint time;
   static const char * keywords[] = {"target_frame", "source_frame", "time", nullptr};
@@ -437,7 +439,7 @@ static PyObject * canTransformCore(PyObject * self, PyObject * args, PyObject * 
 
 static PyObject * canTransformFullCore(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * target_frame, * source_frame, * fixed_frame;
   tf2::TimePoint target_time, source_time;
   static const char * keywords[] =
@@ -477,7 +479,7 @@ static PyObject * asListOfStrings(std::vector<std::string> los)
 
 static PyObject * _chain(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * target_frame, * source_frame, * fixed_frame;
   tf2::TimePoint target_time, source_time;
   std::vector<std::string> output;
@@ -507,7 +509,7 @@ static PyObject * _chain(PyObject * self, PyObject * args, PyObject * kw)
 
 static PyObject * getLatestCommonTime(PyObject * self, PyObject * args)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * target_frame, * source_frame;
   tf2::CompactFrameID target_id, source_id;
   tf2::TimePoint tf2_time;
@@ -578,7 +580,7 @@ cleanup:
 
 static PyObject * lookupTransformCore(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * target_frame, * source_frame;
   tf2::TimePoint time;
   static const char * keywords[] = {"target_frame", "source_frame", "time", nullptr};
@@ -601,7 +603,7 @@ static PyObject * lookupTransformCore(PyObject * self, PyObject * args, PyObject
 
 static PyObject * lookupTransformFullCore(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * target_frame, * source_frame, * fixed_frame;
   tf2::TimePoint target_time, source_time;
   static const char * keywords[] =
@@ -630,7 +632,7 @@ static PyObject * lookupTransformFullCore(PyObject * self, PyObject * args, PyOb
 
 static PyObject * lookupVelocityCore(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = ((buffer_core_t *)self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = ((buffer_core_t *)self)->bc;
   char * tracking_frame, * observation_frame;
   tf2::TimePoint time;
   tf2::Duration averaging_interval;
@@ -658,7 +660,7 @@ static PyObject * lookupVelocityCore(PyObject * self, PyObject * args, PyObject 
 
 static PyObject * lookupVelocityFullCore(PyObject * self, PyObject * args)
 {
-  tf2::BufferCore * bc = ((buffer_core_t *)self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = ((buffer_core_t *)self)->bc;
   char * tracking_frame, * observation_frame, * reference_frame, * reference_point_frame;
   tf2::TimePoint time;
   tf2::Duration averaging_interval;
@@ -737,7 +739,7 @@ static inline int checkRotationType(PyObject * o)
 static PyObject * setTransform(PyObject * self, PyObject * args)
 {
   PyObject * ret = nullptr;
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   PyObject * py_transform;
   char * authority;
   tf2::TimePoint time;
@@ -882,7 +884,7 @@ cleanup:
 static PyObject * setTransformStatic(PyObject * self, PyObject * args)
 {
   PyObject * ret = nullptr;
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   PyObject * py_transform;
   char * authority;
   tf2::TimePoint time;
@@ -1020,14 +1022,14 @@ cleanup:
 static PyObject * clear(PyObject * self, PyObject * args)
 {
   (void)args;
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   bc->clear();
   Py_RETURN_NONE;
 }
 
 static PyObject * _frameExists(PyObject * self, PyObject * args)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   char * frame_id_str;
   if (!PyArg_ParseTuple(args, "s", &frame_id_str)) {
     return nullptr;
@@ -1038,7 +1040,7 @@ static PyObject * _frameExists(PyObject * self, PyObject * args)
 static PyObject * _getFrameStrings(PyObject * self, PyObject * args)
 {
   (void)args;
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   std::vector<std::string> ids;
   bc->_getFrameStrings(ids);
   return asListOfStrings(ids);
@@ -1046,7 +1048,7 @@ static PyObject * _getFrameStrings(PyObject * self, PyObject * args)
 
 static PyObject * _allFramesAsDot(PyObject * self, PyObject * args, PyObject * kw)
 {
-  tf2::BufferCore * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
+  tf2_ros::BufferCoreROSConversions * bc = reinterpret_cast<buffer_core_t *>(self)->bc;
   static const char * keywords[] = {"time", nullptr};
   tf2::TimePoint time;
   if (!PyArg_ParseTupleAndKeywords(
